requirements.txt
discord.py==2.0.0
asyncio
pytz
Pillow
requests